requirements.txt
numpy==1.18.5
pandas==1.0.5
streamlit==0.82.0
Pillow==8.3.0
scikit_learn==0.24.2
